+++ server/zanata-war/src/test/java/org/zanata/service/impl/ActivityServiceImplTest.java
@@ -33,6 +33,7 @@ import com.google.common.base.Throwables;
 
 import org.dbunit.operation.DatabaseOperation;
 import org.hibernate.Session;
+import org.jglue.cdiunit.AdditionalClasses;
 import org.jglue.cdiunit.InRequestScope;
 import org.jglue.cdiunit.deltaspike.SupportDeltaspikeCore;
 import org.junit.Before;
@@ -50,6 +51,7 @@ import org.zanata.events.TextFlowTargetStateEvent;
 import org.zanata.model.Activity;
 import org.zanata.model.type.EntityType;
 import org.zanata.test.CdiUnitRunner;
+import org.zanata.transaction.TransactionUtilImpl;
 import org.zanata.util.IServiceLocator;
 
 import javax.enterprise.inject.Produces;
@@ -62,6 +64,7 @@ import javax.transaction.UserTransaction;
  * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
  */
 @RunWith(CdiUnitRunner.class)
+@AdditionalClasses(TransactionUtilImpl.class)
 @SupportDeltaspikeCore
 public class ActivityServiceImplTest extends ZanataDbunitJpaTest {
