+++ server/zanata-war/src/test/java/org/zanata/transaction/TestTransactionUtil.java
@@ -8,7 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.zanata.util.RunnableEx;
 
-public class TestTransactionUtil implements TXUtil {
+public class TestTransactionUtil implements TransactionUtil {
     private static final Logger log =
             LoggerFactory.getLogger(TestTransactionUtil.class);
     private EntityManager em;
